refactor: AccountConnection is now a container :art:
